Fix so cancel works as expected
Previously, validation errors occur. This errors were generated
on the client side where as we want server side validation. In the
case of adding a new DVD, server side validation had a bug (not
calling the validate() function). Fix of this in turn showed a
bug in the test case for this route. Also turned off CSRF checking
in the test config instead of individual test cases