Fixes for location dependent DVD updates/creations
Now use the enum values directly when checking the existence of and
creating a new DVD.

On the delete form, the select fields need to be string fields as
they are read only. Otherwise, the values are not displayed [BUG].

When pulling location information, needs to be lower cased [BUG].

With the Alembic migration, the server defaults returned are
strings. Thus change the location enum values to be strings.

Need to check for correct DATABASE environment variable to
set a dev environment pointing at PostgreSQL db [BUG].