Add routes for music media LPs
Add templates for listing summary of all LPs and deleting an LP.

Add route displaying a summary listing of all LPs and deleting and LP.

Add find_by_index function to LP music media which is used in the route
to select the LP to delete.

Add api route to gather all LP summary information and supporting
test case.

Fixes:
- comments refering to singleton holding sets corrected to lists
- ensure in "_clean_" functions that the index is reset to 0
- add artist particles getter to music media instances