Implement write on Music Media changes (#115)
Add a flag to the MEDIA super class to indicate changes
have occurred. When a Music Media main page is opened, which
should eventually happen after additions, modifications, and
deletions, write out the html music media file if changes have
occurred. The flag is reset after the changes are written.

Note that this write out is skipped for the test environment
to avoid changing the test music media file.