Fix missing classical composer entries
When entrying song data on a new track, some of the form
ids were incorrect resulting in missing composer data.

The above issue also applied to additional artists.

Also make the labels for composers consistent across all
forms.

The previous fix for duplicate adding of the same music
media item to an artist (composer) erroneously extended
the music media artist list. A copy needed to be created
to check against.