Add new track attributes (Fix #124)
Added track-artist and track-year attributes to the tracklist.
This required updating the templates, Tracklist class, reading
from and writing to the html music media file.

Tests were also updated to include a sample cassette with these
attributes.

Note that the new attributes only apply to cassettes.

Fixed issues with the track mixer to ensure it is an _Artist class
and we are using the name of the track mixer not the _Artist class
in the UI.

Finally add such a cassette to the music media library