Create composite Docker containers with Postgres and the MacMedia app
Compose Docker containers hosting the MacMedia app
and the Postgres database.

The MacMedia app connects to the database which uses
a mounted volume to persist changes to the database.

Secrets are used to pass the database user, password
and database name to both the app and database.

Makefile has been updated to build individual containers and/or
use Docker compose to tie it all together.

Configuration files for Postgres and Postgres hba are included
as they are loaded into the postgres container.